A way to display Independent Candidates
Asking because I can't seem to ascertain this with certainty, but it would be very cool if your app could allow the user to determine independent clients. I am not sure how you'd even define that statistically, but it's the one data point I feel like the app is missing. 

Great work btw. 
-Theena